requirements.txt
discord.py==1.4.1
pendulum
aiohttp
datetime
uuid
time
secrets
pybooru
argparse==1.4.0
python-benedict
Cryptography
termcolor
requests